requirements.txt
paho-mqtt
pyserial
python-dotenv
